UtbetalingsperiodeDetalj til UtbetalingsperiodeDetaljDto før det sendes til frontend
