+++ src/main/kotlin/no/nav/familie/ks/sak/api/dto/BehandlingResponsDto.kt
@@ -12,7 +12,6 @@ import no.nav.familie.ks.sak.kjerne.behandling.steg.VenteÅrsak
 import no.nav.familie.ks.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.Målform
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.PersonType
-import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.YearMonth
@@ -95,19 +94,4 @@ data class YtelsePerioderDto(
     val stønadTom: YearMonth,
     val ytelseType: YtelseType,
     val skalUtbetales: Boolean
-)
-
-data class UtbetalingsperiodeResponsDto(
-    val periodeFom: LocalDate,
-    val periodeTom: LocalDate,
-    val antallBarn: Int,
-    val utbetaltPerMnd: Int,
-    val utbetalingsperiodeDetaljer: List<UtbetalingsperiodeDetaljDto>
-)
-
-data class UtbetalingsperiodeDetaljDto(
-    val person: PersonResponsDto,
-    val utbetaltPerMnd: Int,
-    val erPåvirketAvEndring: Boolean,
-    val prosent: BigDecimal
 )
