+++ src/main/kotlin/no/nav/familie/ks/sak/api/mapper/BehandlingMapper.kt
@@ -8,25 +8,18 @@ import no.nav.familie.ks.sak.api.dto.BehandlingStegTilstandResponsDto
 import no.nav.familie.ks.sak.api.dto.PersonResponsDto
 import no.nav.familie.ks.sak.api.dto.PersonerMedAndelerResponsDto
 import no.nav.familie.ks.sak.api.dto.SøknadDto
-import no.nav.familie.ks.sak.api.dto.UtbetalingsperiodeDetaljDto
 import no.nav.familie.ks.sak.api.dto.UtbetalingsperiodeResponsDto
 import no.nav.familie.ks.sak.api.dto.VedtakDto
 import no.nav.familie.ks.sak.api.dto.YtelsePerioderDto
 import no.nav.familie.ks.sak.api.mapper.RegisterHistorikkMapper.lagRegisterHistorikkResponsDto
-import no.nav.familie.ks.sak.common.exception.Feil
-import no.nav.familie.ks.sak.common.tidslinje.filtrerIkkeNull
-import no.nav.familie.ks.sak.common.tidslinje.utvidelser.tilPerioder
 import no.nav.familie.ks.sak.common.util.toYearMonth
 import no.nav.familie.ks.sak.kjerne.arbeidsfordeling.domene.ArbeidsfordelingPåBehandling
 import no.nav.familie.ks.sak.kjerne.behandling.domene.Behandling
 import no.nav.familie.ks.sak.kjerne.behandling.steg.BehandlingStegStatus
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.PersonResultat
-import no.nav.familie.ks.sak.kjerne.beregning.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ks.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ks.sak.kjerne.beregning.domene.slåSammenBack2BackAndelsperioderMedSammeBeløp
-import no.nav.familie.ks.sak.kjerne.beregning.lagVertikalePerioder
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.Person
-import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.PersonopplysningGrunnlag
 import java.math.BigDecimal
 import java.time.LocalDate
 
@@ -122,39 +115,4 @@ object BehandlingMapper {
                     }
                 )
             }
-
-    fun lagUtbetalingsperioder(
-        personopplysningGrunnlag: PersonopplysningGrunnlag,
-        andelerTilkjentYtelseMedEndreteUtbetalinger: List<AndelTilkjentYtelseMedEndreteUtbetalinger>
-    ): List<UtbetalingsperiodeResponsDto> {
-        if (andelerTilkjentYtelseMedEndreteUtbetalinger.isEmpty()) return emptyList()
-        return andelerTilkjentYtelseMedEndreteUtbetalinger.lagVertikalePerioder().tilPerioder().filtrerIkkeNull().map {
-            val andelerForPeriode = it.verdi
-            val sumUtbetalingsbeløp = andelerForPeriode.sumOf { andel -> andel.kalkulertUtbetalingsbeløp }
-            UtbetalingsperiodeResponsDto(
-                periodeFom = checkNotNull(it.fom),
-                periodeTom = checkNotNull(it.tom),
-                utbetaltPerMnd = sumUtbetalingsbeløp,
-                antallBarn = andelerForPeriode.count { andel ->
-                    personopplysningGrunnlag.barna.any { barn -> barn.aktør == andel.aktør }
-                },
-                utbetalingsperiodeDetaljer = andelerForPeriode.lagUtbetalingsperiodeDetaljer(personopplysningGrunnlag)
-            )
-        }
-    }
-
-    private fun Collection<AndelTilkjentYtelseMedEndreteUtbetalinger>.lagUtbetalingsperiodeDetaljer(
-        personopplysningGrunnlag: PersonopplysningGrunnlag
-    ): List<UtbetalingsperiodeDetaljDto> =
-        this.map { andel ->
-            val personForAndel = personopplysningGrunnlag.personer.find { person -> andel.aktør == person.aktør }
-                ?: throw Feil("Fant ikke personopplysningsgrunnlag for andel")
-
-            UtbetalingsperiodeDetaljDto(
-                person = lagPersonRespons(personForAndel, emptyMap()),
-                utbetaltPerMnd = andel.kalkulertUtbetalingsbeløp,
-                erPåvirketAvEndring = andel.endreteUtbetalinger.isNotEmpty(),
-                prosent = andel.prosent
-            )
-        }
 }
