+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vedtak/vedtaksperiode/Utbetalingsperiode.kt
@@ -10,8 +10,7 @@ import no.nav.familie.ks.sak.common.util.sisteDagIInneværendeMåned
 import no.nav.familie.ks.sak.kjerne.beregning.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ks.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ks.sak.kjerne.beregning.tilKombinertTidslinjePerAktør
-import no.nav.familie.ks.sak.kjerne.brev.domene.BrevPerson
-import no.nav.familie.ks.sak.kjerne.brev.domene.tilBrevPerson
+import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.Person
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.PersonopplysningGrunnlag
 import java.math.BigDecimal
 import java.time.LocalDate
@@ -31,7 +30,7 @@ data class Utbetalingsperiode(
 ) : Vedtaksperiode
 
 data class UtbetalingsperiodeDetalj(
-    val brevPerson: BrevPerson,
+    val brevPerson: Person,
     val ytelseType: YtelseType,
     val utbetaltPerMnd: Int,
     val erPåvirketAvEndring: Boolean,
@@ -42,7 +41,6 @@ fun mapTilUtbetalingsperioder(
     personopplysningGrunnlag: PersonopplysningGrunnlag,
     andelerTilkjentYtelse: List<AndelTilkjentYtelseMedEndreteUtbetalinger>
 ): List<Utbetalingsperiode> {
-
     val kombinertTidslinjePerAktør = andelerTilkjentYtelse.tilKombinertTidslinjePerAktør()
 
     val utbetalingsPerioder = kombinertTidslinjePerAktør.tilPerioderIkkeNull().map {
@@ -61,7 +59,9 @@ fun mapTilUtbetalingsperioder(
 
 private fun List<AndelTilkjentYtelseMedEndreteUtbetalinger>.tilTidslinje() = map {
     Periode(
-        it, it.stønadFom.førsteDagIInneværendeMåned(), it.stønadTom.sisteDagIInneværendeMåned()
+        it,
+        it.stønadFom.førsteDagIInneværendeMåned(),
+        it.stønadTom.sisteDagIInneværendeMåned()
     )
 }.tilTidslinje()
 
@@ -74,7 +74,7 @@ internal fun Collection<AndelTilkjentYtelseMedEndreteUtbetalinger>.lagUtbetaling
         )
 
     UtbetalingsperiodeDetalj(
-        brevPerson = personForAndel.tilBrevPerson(),
+        brevPerson = personForAndel,
         ytelseType = andel.type,
         utbetaltPerMnd = andel.kalkulertUtbetalingsbeløp,
         erPåvirketAvEndring = andel.endreteUtbetalinger.isNotEmpty(),
