+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vedtak/vedtaksperiode/domene/VedtaksperiodeService.kt
@@ -333,7 +333,7 @@ class VedtaksperiodeService(
     ): List<Aktør> = utvidetVedtaksperiodeMedBegrunnelser
         .utbetalingsperiodeDetaljer
         .map { utbetalingsperiodeDetalj ->
-            val ident = utbetalingsperiodeDetalj.brevPerson.aktivPersonIdent
+            val ident = utbetalingsperiodeDetalj.brevPerson.aktør.aktivFødselsnummer()
             persongrunnlag.personer.find { it.aktør.aktivFødselsnummer() == ident }?.aktør
                 ?: personidentService.hentAktør(ident)
         }
